RSDEV-1246-template-links-with-targets-set: fix message keys, lift duplicated link helpers
Fix, found while reviewing the rebase: the instrument template's two
field-delete errors were still throwing this branch's original message keys,
errors.inventory.field.deleteRequest.{idMissing,idUnknown}. Main externalised
the same two messages itself during the JSON-catalogue migration, flat, as
deleteRequestId{Missing,Unknown}, so those keys exist in no catalogue and both
error paths resolved to nothing. Compiles, lints and passes i18n:lint either
way; it only shows against a live server. Guarded by a test that takes the code
from the production path and asserts it resolves through JsonMessageSource,
rather than restating the literal.

Complexity, from ponytail-review: the link write path was duplicated almost
verbatim across the sample and instrument managers. applyLinkFieldValue's two
42-line bodies were identical apart from comments, as were assertRelationAllowed
and parseTargetOrNull; rejectSelfLink differed only in the entity type, which
InventoryRecord already covers. All of them, plus applyLinkFieldValuesOnUpdate
and applyDefaultLinkOfNewTemplateField, move to InventoryApiManagerImpl, which
already hosts inventoryLinkManager, isRelationPermitted,
assertDefaultLinksMatchWhitelists and softDeleteLinkOfDeletedLinkField. Each
manager keeps a short typed delegate for applyLinkFieldValuesOnUpdate, because
isTemplate() lives on the sample and instrument entities rather than on
InventoryRecord. Net -131 lines of Java.

Also in LinkEditor: drop the IIFE that existed only to bind the target icon, and
hoist the group-heading sx written twice.
